python/requirements.txt
requests
flask
python-jenkins
waitress
